+++ internal/app/handlers/health_checks.go
@@ -16,6 +16,7 @@ type HealthChecksHandler struct {
 	webAppsStore      types.WebAppsStore
 	subscribersStore  types.SubscribersStore
 	healthChecksStore types.HealthChecksStore
+	emailClient       types.EmailClient
 }
 
 func NewHealthChecksHandler(
@@ -23,12 +24,14 @@ func NewHealthChecksHandler(
 	was types.WebAppsStore,
 	sus types.SubscribersStore,
 	hcs types.HealthChecksStore,
+	emc types.EmailClient,
 ) *HealthChecksHandler {
 	return &HealthChecksHandler{
 		settings:          stg,
 		webAppsStore:      was,
 		subscribersStore:  sus,
 		healthChecksStore: hcs,
+		emailClient:       emc,
 	}
 }
 
@@ -47,7 +50,7 @@ func (h *HealthChecksHandler) GetHealthChecks(res http.ResponseWriter, req *http
 		return
 	}
 
-	page, err := cmd.NewGetHealthChecksCmd(h.healthChecksStore).Execute(id, req.URL.Query())
+	page, err := cmd.NewGetHealthChecksCmd(h.healthChecksStore).Execute(app.Id, req.URL.Query())
 
 	if err != nil {
 		helpers.InternalServerError(err, res, req)
@@ -58,8 +61,39 @@ func (h *HealthChecksHandler) GetHealthChecks(res http.ResponseWriter, req *http
 	helpers.Render(template, res, req)
 }
 
+func (h *HealthChecksHandler) GetDoughnutHealthChart(res http.ResponseWriter, req *http.Request) {
+	user := helpers.AuthPrincipal(req)
+	id := req.PathValue("id")
+	app, _, err := cmd.NewGetWebAppCmd(h.webAppsStore, h.subscribersStore).Execute(id, user)
+
+	if err != nil {
+		if err == sql.ErrNoRows {
+			helpers.NotFoundError(res, req)
+			return
+		}
+
+		helpers.InternalServerError(err, res, req)
+		return
+	}
+
+	health, err := cmd.NewGetDoughnutHealthChartCmd(h.healthChecksStore).Execute(app)
+
+	if err != nil {
+		helpers.InternalServerError(err, res, req)
+		return
+	}
+
+	template := components.DoughnutHealthChart(health)
+	helpers.Render(template, res, req)
+}
+
 func (h *HealthChecksHandler) PeriodicHealthCheck(itv string) {
-	cmd.NewPeriodicHealthCheckCmd(h.webAppsStore, h.healthChecksStore).Execute(itv)
+	cmd.NewPeriodicHealthCheckCmd(
+		h.webAppsStore,
+		h.subscribersStore,
+		h.healthChecksStore,
+		h.emailClient,
+	).Execute(itv)
 }
 
 func (h *HealthChecksHandler) HealthChecksRetention() {