+++ internal/app/types/clients.go
@@ -1,0 +1,5 @@
+package types
+
+type EmailClient interface {
+	SendEmail(to, sub, msg string) error
+}